refactor(ios): split SimulatorBridge (1964 LOC) en 4 archivos focalizados
Deuda técnica post-ARD-001 (PR #106) cerrada. SimulatorBridge.swift era
un monolito de 1964 LOC con 6 responsabilidades mezcladas — complicaba
reading, testing y evolución. Ahora está físicamente extraído en 4
archivos usando extensions del mismo tipo (cero cambios a la API pública).

Estrategia: **extension-based split**. Cada "engine" es una extension de
`SimulatorBridge` en un archivo separado. Mantiene acceso a stored
properties y preserva la interface pública — los helpers privados
compartidos entre archivos se promovieron a internal (default
visibility). Ninguna fase rompe la API ni cambia comportamiento.

Layout final:

  SimulatorBridge.swift                    677 LOC  (core: gestures + input + state)
    - State: simulatorPID, recordingProcess, recordingTempPath
    - Gestures: tap, longPress, doubleTap, clear, typeText, scroll,
      swipe, drag, tapAtCoordinate
    - Input helpers: trySetFocusedFieldValue, ensureHardwareKeyboardEnabled,
      pasteViaPasteboard, autoDismissPasteboardDialog, typeCharByChar,
      keyCodeForChar
    - Internals: click, resolveElementCenter, logTap
    - Extension DeviceBridge: tree(), copyTextFrom

  SimulatorBridge+AXEngine.swift           521 LOC  (AX queries + element search)
    - findSimulator, findSimulatorPID, findSimulatorContent/Fast,
      getSimulatorWindowFrame
    - tree, search, elementAt, existsFast
    - findAXElement (+ Exact/Contains), findElementInfo, findElementAt
    - findAXElementScoped (con role + within + label[N] fallbacks)
    - tapElement
    - serializeElement, serializeChildren, searchRecursive
    - AX helpers: getAttribute, getChildren, getPosition, getSize,
      getFirstWindow, getParent, getRole

  SimulatorBridge+SimCtlEngine.swift       677 LOC  (simctl + biometry + files)
    - getBootedDeviceId, listDevices, bootDevice, shutdownDevice
    - launchApp, terminateApp, installApp, uninstallApp, clearState
    - openURL, setPasteboard, getPasteboard
    - Biometric: biometricEnroll/Unenroll/Match/Fail/IsEnrolled + faceID aliases
    - getLogs, setPermission, resetKeychain
    - pressKey, hideKeyboard, eraseText
    - rotate, lockDevice, unlockDevice
    - setLocation, setAppearance, viewport
    - pushFile, pullFile
    - Helpers: activateSimulatorApp, clickSimulatorMenu, resolveDeviceUdid,
      toggleBiometricEnrollment

  SimulatorBridge+MediaEngine.swift        200 LOC  (screenshot + recording + camera)
    - screenshot, addMedia
    - startRecording, stopRecording
    - cameraStart/Feed/Stop/Status (+ cameraFeedPath, cameraSignalPath)
    - buildWithCameraMock, setInjectImage, injectAndLaunch
    - injectImagePath (static)

  Total: 2075 LOC (vs 1964 original — +5% por comentarios + imports
  en cada archivo). SimulatorBridge.swift: 1964 → 677 LOC (-65%).

Invariantes preservados (verificado):
  * API pública de SimulatorBridge — sin cambios (zero impact en CLI,
    RecordingSession, LegacyBridgeAdapter, backends ARD-001)
  * LegacyBridgeAdapter(simulatorBridge) sigue funcionando idéntico
  * swift build — 0 warnings, 0 errors
  * swift test — 121/121 verdes (sin regresiones)
  * auto run smoke-ard001.auto iOS: 7/7 en 5218ms
  * auto-android run smoke-ard001.auto: 7/7 en 2026ms
