+++ cli/Sources/AutoLibiOS/SimulatorBridge+MediaEngine.swift
@@ -1,0 +1,200 @@
+import Foundation
+import AutoCore
+
+// MARK: - MediaEngine
+//
+// Captura de media y camera mock iOS. Extraído de `SimulatorBridge.swift` como
+// parte de la "extracción física" post-ARD-001 — la API pública del bridge no
+// cambia, el código solo vive en este archivo separado para facilitar reading
+// y testing focalizado.
+//
+// Responsabilidades:
+//   - `screenshot` / `addMedia` vía `xcrun simctl io`
+//   - `startRecording` / `stopRecording` con state en SimulatorBridge
+//   - `cameraStart/Feed/Stop/Status` vía filesystem (feed + signal files)
+//   - `buildWithCameraMock` / `setInjectImage` / `injectAndLaunch` (DYLD inject)
+
+extension SimulatorBridge {
+
+    // MARK: Paths compartidos (fileprivate visibles en esta extension)
+
+    /// Well-known path donde el mock dylib lee imágenes.
+    public static let injectImagePath = "/tmp/autopilot-camera-image.jpg"
+
+    internal static let cameraFeedPath = "/tmp/autopilot-camera-feed.jpg"
+    internal static let cameraSignalPath = "/tmp/autopilot-camera-active"
+
+    // MARK: - Screenshot
+
+    public func screenshot(path: String) throws {
+        let deviceId = try getBootedDeviceId()
+        let process = Process()
+        process.executableURL = URL(fileURLWithPath: "/usr/bin/xcrun")
+        process.arguments = ["simctl", "io", deviceId, "screenshot", path]
+        try process.run()
+        process.waitUntilExit()
+        guard process.terminationStatus == 0 else {
+            throw BridgeError.screenshotFailed
+        }
+    }
+
+    // MARK: - Add Media (simctl addmedia)
+
+    public func addMedia(path: String) throws {
+        let deviceId = try getBootedDeviceId()
+        let process = Process()
+        process.executableURL = URL(fileURLWithPath: "/usr/bin/xcrun")
+        process.arguments = ["simctl", "addmedia", deviceId, path]
+        try process.run()
+        process.waitUntilExit()
+        guard process.terminationStatus == 0 else {
+            throw BridgeError.mediaInjectionFailed(path)
+        }
+    }
+
+    // MARK: - Screen Recording
+
+    public func startRecording() throws {
+        let deviceId = try getBootedDeviceId()
+        let tempPath = NSTemporaryDirectory() + "autopilot-recording-\(ProcessInfo.processInfo.globallyUniqueString).mp4"
+        recordingTempPath = tempPath
+
+        let process = Process()
+        process.executableURL = URL(fileURLWithPath: "/usr/bin/xcrun")
+        process.arguments = ["simctl", "io", deviceId, "recordVideo", "--codec=h264", tempPath]
+        process.standardOutput = Pipe()
+        process.standardError = Pipe()
+        try process.run()
+        recordingProcess = process
+    }
+
+    public func stopRecording(outputPath: String) throws {
+        guard let process = recordingProcess else {
+            throw BridgeError.unknown("No recording in progress")
+        }
+
+        // Send SIGINT to stop recording gracefully
+        process.interrupt()
+        process.waitUntilExit()
+
+        guard let tempPath = recordingTempPath else {
+            throw BridgeError.unknown("No recording temp path found")
+        }
+
+        let fm = FileManager.default
+        if fm.fileExists(atPath: outputPath) {
+            try fm.removeItem(atPath: outputPath)
+        }
+        try fm.moveItem(atPath: tempPath, toPath: outputPath)
+
+        recordingProcess = nil
+        recordingTempPath = nil
+    }
+
+    // MARK: - Virtual Camera (feed + signal files)
+
+    /// Start virtual camera with an image.
+    public func cameraStart(imagePath: String) throws {
+        let source = URL(fileURLWithPath: imagePath)
+        guard FileManager.default.fileExists(atPath: imagePath) else {
+            throw BridgeError.cameraImageNotFound(imagePath)
+        }
+
+        let dest = URL(fileURLWithPath: Self.cameraFeedPath)
+        if FileManager.default.fileExists(atPath: Self.cameraFeedPath) {
+            try FileManager.default.removeItem(at: dest)
+        }
+        try FileManager.default.copyItem(at: source, to: dest)
+
+        // Crear archivo signal
+        FileManager.default.createFile(atPath: Self.cameraSignalPath, contents: nil)
+    }
+
+    /// Update the camera feed image.
+    public func cameraFeed(imagePath: String) throws {
+        guard FileManager.default.fileExists(atPath: imagePath) else {
+            throw BridgeError.cameraImageNotFound(imagePath)
+        }
+
+        let dest = URL(fileURLWithPath: Self.cameraFeedPath)
+        if FileManager.default.fileExists(atPath: Self.cameraFeedPath) {
+            try FileManager.default.removeItem(at: dest)
+        }
+        try FileManager.default.copyItem(at: URL(fileURLWithPath: imagePath), to: dest)
+    }
+
+    /// Stop the virtual camera.
+    public func cameraStop() {
+        try? FileManager.default.removeItem(atPath: Self.cameraSignalPath)
+        try? FileManager.default.removeItem(atPath: Self.cameraFeedPath)
+    }
+
+    /// Check if virtual camera is active.
+    public func cameraStatus() -> (active: Bool, imagePath: String?) {
+        let active = FileManager.default.fileExists(atPath: Self.cameraSignalPath)
+        let hasImage = FileManager.default.fileExists(atPath: Self.cameraFeedPath)
+        return (active: active, imagePath: hasImage ? Self.cameraFeedPath : nil)
+    }
+
+    // MARK: - Build (xcodebuild + camera mock VFS overlay)
+
+    /// Builds an Xcode project with camera mock injected via VFS overlay.
+    public func buildWithCameraMock(args: [String]) throws {
+        let interceptor = BuildInterceptor()
+        try interceptor.build(args: args)
+    }
+
+    // MARK: - Inject (dylib — DYLD_INSERT_LIBRARIES)
+
+    /// Copies an image to the well-known path so the mock picks it up on next capture.
+    public func setInjectImage(_ sourcePath: String) throws {
+        let absPath = sourcePath.hasPrefix("/") ? sourcePath : FileManager.default.currentDirectoryPath + "/" + sourcePath
+        guard FileManager.default.fileExists(atPath: absPath) else {
+            throw InjectError.imageNotFound(absPath)
+        }
+        let dest = SimulatorBridge.injectImagePath
+        try? FileManager.default.removeItem(atPath: dest)
+        try FileManager.default.copyItem(atPath: absPath, toPath: dest)
+    }
+
+    /// Launches an app with the camera mock dylib injected via DYLD_INSERT_LIBRARIES.
+    /// Optionally sets an initial image for the mock.
+    public func injectAndLaunch(bundleId: String, imagePath: String?, extraEnv: [String: String] = [:]) throws {
+        let injector = DylibInjector()
+        let dylibPath = try injector.ensureDylib()
+
+        // Copy initial image if provided
+        if let img = imagePath {
+            try setInjectImage(img)
+        }
+
+        let deviceId = try getBootedDeviceId()
+
+        // Terminate if running (ignore errors — app may not be running)
+        let term = Process()
+        term.executableURL = URL(fileURLWithPath: "/usr/bin/xcrun")
+        term.arguments = ["simctl", "terminate", deviceId, bundleId]
+        try? term.run()
+        term.waitUntilExit()
+
+        // Build env vars — only need dylib path, image is read from fixed file
+        var env = ProcessInfo.processInfo.environment
+        env["SIMCTL_CHILD_DYLD_INSERT_LIBRARIES"] = dylibPath
+
+        for (key, value) in extraEnv {
+            env["SIMCTL_CHILD_\(key)"] = value
+        }
+
+        // Launch with dylib injection
+        let launch = Process()
+        launch.executableURL = URL(fileURLWithPath: "/usr/bin/xcrun")
+        launch.arguments = ["simctl", "launch", deviceId, bundleId]
+        launch.environment = env
+        try launch.run()
+        launch.waitUntilExit()
+
+        guard launch.terminationStatus == 0 else {
+            throw BridgeError.simctlFailed("simctl launch failed (exit \(launch.terminationStatus))")
+        }
+    }
+}
